requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.12.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.12
django-celery==3.3.1
django-cors-headers==3.2.0
django-environ==0.4.5
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
drf-yasg==1.17.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==3.0.37
MarkupSafe==1.1.1
openpyxl==3.0.5
packaging==20.7
psycopg2-binary==2.8.4
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.1.0
